backend/requirements.txt
Flask
pymongo
flask-jwt-extended
werkzeug
imapclient
